refactor: 로또 번호 발행을 담당하는 기능들은 Vender 클래스로 분리
